% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postProcessing.R
\name{add_metadata_to_metabolite_summary}
\alias{add_metadata_to_metabolite_summary}
\title{Do basic enrichment testing of between-group metabolite differences and add results to the MIMOSA summary table}
\usage{
add_metadata_to_metabolite_summary(node_data, mets, metadata, metadata_var)
}
\arguments{
\item{node_data}{Table from "_nodes.txt" file produced by runMimosa.R}

\item{mets}{Table of metabolite abundance data}

\item{metadata}{Table of samples and their grouping by a binary variable}

\item{metadata_var}{Column name in metadata table for the binary grouping variable to be used for testing}
}
\value{
Processed table including columns describing binary enrichment status and difference in mean
add_metadata_to_metabolite_summary(node_data, mets, metadata, "DiseaseStatus")
}
\description{
Do basic enrichment testing of between-group metabolite differences and add results to the MIMOSA summary table
}
