% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helperMethods.R
\name{asinh_trans}
\alias{asinh_trans}
\title{asinh_trans}
\usage{
asinh_trans(c)
}
\arguments{
\item{c}{\code{numeric} cofactor for asinh trasnform. Default 1.}
}
\value{
transform
}
\description{
Arcsinh transform for ggplot2
}
\details{
Arcsinh transform for use with coord_trans in ggplot2
}
\author{
Greg Finak
}

