% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{BeetlesBody}
\alias{BeetlesBody}
\title{Simulated beetle body-length data}
\format{
## `BeetlesBody`
A data frame with 960 rows and 6 columns:
\describe{
  \item{Population}{Population identifier.}
  \item{Container}{Container identifier.}
  \item{Sex}{Beetle sex: `Female` or `Male`.}
  \item{Habitat}{Microhabitat: `A` or `B`.}
  \item{Treatment}{Dietary treatment: `Cont` or `Exp`.}
  \item{BodyL}{Body length.}
}
}
\source{
<https://github.com/mastoffel/rptR/blob/master/data/BeetlesBody.rda>
}
\usage{
data(BeetlesBody)
}
\description{
A balanced simulated data set from the worked example in Nakagawa and
Schielzeth (2013). Its imaginary sampling design contains beetles from 12
populations and 120 containers, with two habitats, two dietary treatments,
and both sexes. Body length is the continuous response.
}
\references{
Nakagawa, S. & Schielzeth, H. (2013). A general and simple method for
obtaining R2 from generalized linear mixed-effects models. *Methods in
Ecology and Evolution*, 4, 133–142. \doi{10.1111/j.2041-210x.2012.00261.x}
}
\keyword{datasets}
