% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperMethods.R
\name{boxplotMIMOSAResultList}
\alias{boxplotMIMOSAResultList}
\title{boxplotMIMOSAResultList}
\usage{
boxplotMIMOSAResultList(
  data,
  title = "A Boxplot",
  x_axis_category = NULL,
  cofactor = 5000,
  line = TRUE,
  threshold = 0.005
)
}
\arguments{
\item{data}{\code{MIMOSAResultList}}

\item{title}{\code{character} Title of the plot.}

\item{x_axis_category}{\code{name} the column of the phenoData frame for the x-axis of the boxplots.}

\item{cofactor}{\code{integer} cofactor of the arcsinhTransform for the y axis.}

\item{line}{\code{logical} whether or not to connect points from the same subject}

\item{threshold}{\code{numeric} the FDR threshold (q-value) at which to classify responders as a separate category.}
}
\value{
\code{ggplot} object.
}
\description{
Boxplots of MIMOSA
}
\details{
Generate boxplots for MIMOSA positivity calls.
}
\author{
Greg Finak
}
