% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mimosa.R
\name{cat_function}
\alias{cat_function}
\title{function to install packages . If they exist load them else go to (CRAN or github) and install them}
\usage{
cat_function(test_level = TRUE)
}
\arguments{
\item{test_level}{Do you want to be in first block or second block? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{block}
\keyword{print}
