% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleSpec_contributions.R
\name{cmp_species_contributions_picrust}
\alias{cmp_species_contributions_picrust}
\title{Evaluate species contributors for a single metabolite with OTU and PICRUSt data}
\usage{
cmp_species_contributions_picrust(j, cmps_sub_good, all_rxns, subjects,
  norm_kos, ko_net, all_taxa, single_spec_cmps, cor_with = F)
}
\arguments{
\item{j}{metabolite # (usually from lapply/sapply)}

\item{cmps_sub_good}{CMP scores for metabolites with abundance data}

\item{all_rxns}{list of relevant reactions for each metabolite}

\item{subjects}{vector of subjects}

\item{norm_kos}{data.table of gene abundances}

\item{ko_net}{network created by generate_genomic_network}

\item{all_taxa}{vector of OTUs}

\item{single_spec_cmps}{single-species CMP scores calculated from get_spec_contribs function}

\item{cor_with}{whether to look at the correlation of CMP scores of each species by itself with the metabolite, or of the whole community with that species removed}
}
\value{
list of 2-item lists for every metabolite - 1st item is data.table of OTUs and correlations, second item is vector of OTUs with correlations > 0.5
}
\description{
Evaluate species contributors for a single metabolite with OTU and PICRUSt data
}
\examples{
sapply(1:length(metabolites), cmp_species_contributions_picrust, cmp_scores, all_rxns,
all_subjects, ko_abunds, ko_net, spec_abunds, ref_kos)
}

