% Generated by roxygen2 (4.0.2): do not edit by hand
\name{combine.MIMOSA}
\alias{combine.MIMOSA}
\title{Combine MIMOSAResultList objects}
\usage{
combine.MIMOSA(x, y, ...)
}
\arguments{
\item{x}{\code{MIMOSAResultList}}

\item{y}{\code{MIMOSAResultList}}

\item{...}{additional \code{MIMOSAResultList} objects}
}
\value{
a \code{MIMOSAResultList}
}
\description{
Combine two or more MIMOSAResultList objects
}
\details{
Combines two or more MIMOSAResultList objects. The method is light on error checking so the results should be from the same
MIMOSAExpressionSet object.
}
\author{
Greg Finak
}

