% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{compare_met}
\alias{compare_met}
\title{Compare a single set each of CMP scores and metabolite concentrations}
\usage{
compare_met(met_met, met_cmp, met_all, cmp_all, posneg = "pos",
  cor_method = "spearman", nperm = 15000)
}
\arguments{
\item{met_met}{metabolite to use from metabolite concentration data}

\item{met_cmp}{metabolite to use from CMP score data}

\item{met_all}{matrix of metabolite concentrations}

\item{cmp_all}{matrix of CMP scores}

\item{posneg}{Whether to test for positive or negative concentration}

\item{cor_method}{spearman or pearson}

\item{nperm}{number of permutations for Mantel test}
}
\value{
p-value from Mantel test
}
\description{
Compare a single set each of CMP scores and metabolite concentrations
}
\examples{
compare_met("C00001", "C00002", mets, cmp_mat, "pos")
}
