% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{correct}
\alias{correct}
\title{Multiple hypothesis correction}
\usage{
correct(pvals, method = "fdr")
}
\arguments{
\item{pvals}{Vector of p-values}

\item{method}{Must be either "fdr" or "bonferroni"}
}
\value{
Vector of corrected values
}
\description{
Multiple hypothesis correction
}
\examples{

}

