% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_stats.R
\name{count_stats}
\alias{count_stats}
\title{Calculate performance statistics}
\usage{
count_stats(gold_standard, predicted_segmentation, k, percent_overlap = NULL,
  verbose = TRUE)
}
\arguments{
\item{gold_standard}{Gold standard segmentation mask of class nifti}

\item{predicted_segmentation}{Predicted segmentation mask volume of class nifti}

\item{k}{Minimum number of voxels for a segmentation cluster/component}

\item{percent_overlap}{Proportion of gold standard segmentation to be overlaped by predicted}

\item{verbose}{Logical indicating printing diagnostic output}
}
\value{
List of objects
}
\description{
This function calculates true positive rate, false positive rate, false negative rate, false positive count, and sensitivity
}
\examples{
\dontrun{

}
}
