\name{fdrComparison}
\alias{fdrComparison}
\title{fdrComparison calculates the observed vs expected false discovery rate}
\usage{
  fdrComparison(fdr, truth)
}
\arguments{
  \item{fdr}{is the expected false discovery rate}

  \item{truth}{is a logical with the true positive and
  negative results}
}
\description{
  fdrComparison calculates the observed vs expected false
  discovery rate
}

