% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{find_id}
\alias{find_id}
\title{Tries to find the grouping variable (id)}
\usage{
find_id(d)
}
\arguments{
\item{d}{data frame}
}
\value{
potential grouping variables
}
\description{
Ranks candidate grouping variables using diagnostics from
\code{explain_find_id()}.
}
\examples{
find_id(mlmRev::Exam)
}
