% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{gene_contributions}
\alias{gene_contributions}
\title{Identify potential important gene contributors for each metabolite.}
\usage{
gene_contributions(j, prmts_sub_good, all_rxns, subjects, norm_kos, ko_net)
}
\arguments{
\item{j}{index (from lapply usually)}

\item{prmts_sub_good}{CMP scores for metabolites to be analyzed}

\item{all_rxns}{list of reaction tables for each compound}

\item{subjects}{vector of subject names}

\item{norm_kos}{gene abundance matrix}

\item{ko_net}{full network}
}
\value{
list of contributors and correlations
}
\description{
Identify potential important gene contributors for each metabolite.
}
\examples{
lapply(1:length(good_mets), prmt_contributions, prmts_sub_good = prmts_sub_good, all_rxns = all_rxns[[j]],
subjects=subjects, norm_kos = norm_kos, ko_net = ko_net)
}

