% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{generate_network_template_kegg}
\alias{generate_network_template_kegg}
\title{Generate community metabolic network template from the KEGG database, using reaction_mapformula.lst and}
\usage{
generate_network_template_kegg(mapformula_file, all_kegg)
}
\arguments{
\item{mapformula_file}{file path to "reaction_mapformula.lst" from the KEGG database, specifying reaction IDs annotated in pathways}

\item{all_kegg}{Output of get_kegg_reaction_info}
}
\value{
A table specifying a community metabolic network based on KEGG pathways with the following columns: Rxn (reaction ID),	KO (Gene ID),	Reac (Reactant ID),	Prod (Product ID),	Path (Pathway ID),	ReacProd (Original reaction specification),	stoichReac (Reactant stoichiometry coefficient),	stoichProd (Product stoichiometry coefficient)
}
\description{
Generate community metabolic network template from the KEGG database, using reaction_mapformula.lst and
}
\examples{
generate_network_template_kegg("KEGG/ligand/reaction/reaction_mapformula.lst", "KEGGREST")
generate_network_template_kegg("KEGG/ligand/reaction/reaction_mapformula.lst",
"KEGG/genes/ko/ko_reaction.list", "KEGG/ligand/reaction/reaction")
}

