% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postProcessing.R
\name{get_all_consistent_metabolite_info}
\alias{get_all_consistent_metabolite_info}
\title{Read output of runMimosa.R from files to make a table of all relevant MIMOSA results for metabolites identified as consistent}
\usage{
get_all_consistent_metabolite_info(prefix, contribs_file)
}
\arguments{
\item{prefix}{Naming prefix used by runMimosa.R}

\item{contribs_file}{File of species-specific gene abundances used for calculating species contributions}
}
\value{
A table where every row represents genes and species that were identified as potential important contributors for a consistent metabolite
get_all_consistent_metabolite_info("run1", "filePath/otu_gene_contributions.txt")
}
\description{
Read output of runMimosa.R from files to make a table of all relevant MIMOSA results for metabolites identified as consistent
}
