% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleSpec_contributions.R
\name{get_all_singleSpec_cmps}
\alias{get_all_singleSpec_cmps}
\title{Get single-species CMP scores for every OTU and metabolite}
\usage{
get_all_singleSpec_cmps(all_otus, all_koAbunds_byOTU, valueVar, out_prefix,
  rxn_table, degree_filter = 30, write_out = T)
}
\arguments{
\item{all_otus}{vector of OTUs or taxa}

\item{all_koAbunds_byOTU}{list of KO abundances for each OTU}

\item{valueVar}{"relAbundSample" or "singleMusicc", abundance metric to use for single taxon gene abundances}

\item{out_prefix}{Path of directory for loading and saving output}

\item{rxn_table}{Community network template}

\item{degree_filter}{Filter compounds connected to more than this number of KOs in the metabolic network model (default 30)}

\item{write_out}{Whether to save output}
}
\value{
list of matrices of cmp scores for each taxon
}
\description{
Get single-species CMP scores for every OTU and metabolite
}

