% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleSpec_contributions.R
\name{get_all_singleSpec_prmts}
\alias{get_all_singleSpec_prmts}
\title{Get single-species CMP scores for every OTU and metabolite}
\usage{
get_all_singleSpec_prmts(all_otus, valueVar)
}
\arguments{
\item{all_otus}{vector of OTUs or taxa}

\item{valueVar}{"relAbundSample" or "singleMusicc", abundance metric to use for single taxon gene abundances}
}
\value{
Null, saves to Rdata file
}
\description{
Get single-species CMP scores for every OTU and metabolite
}

