% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helper_functions.R
\name{get_net_dist}
\alias{get_net_dist}
\title{Get network distances between 2 compounds}
\usage{
get_net_dist(c1, c2, allnet, max_dist = 20)
}
\arguments{
\item{c1}{first compound}

\item{c2}{second compound}

\item{allnet}{edge list output of generate_genomic_network}

\item{max_dist}{distance at which the function will stop looking and assume the two compounds are not connected}
}
\value{
integer representing the number of reaction steps between the two compounds, or max_dist if they are not connected
}
\description{
Get network distances between 2 compounds
}
\examples{
get_net_dist("C00001", "C00002", allnet)
}
