% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helper_functions.R
\name{get_non_rev_rxns}
\alias{get_non_rev_rxns}
\title{Get non-reverible reactions of network.}
\usage{
get_non_rev_rxns(rxn_table, all_rxns = T, by_species = T)
}
\arguments{
\item{rxn_table}{Edge list}

\item{all_rxns}{Whether to return all reactions or only 1/2 of each reversible reaction since info is redundant}

\item{by_species}{Whether this is a species-specific network}
}
\value{
Network with reversibility informatio
}
\description{
Get non-reverible reactions of network.
}
\examples{
get_non_rev_rxns(rxn_table)
}
