% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{get_prmt_scores}
\alias{get_prmt_scores}
\title{Calculate CMP scores based on community network and gene abundances}
\usage{
get_prmt_scores(emm, norm_kos)
}
\arguments{
\item{emm}{Stoichiometric network matrix produced by generate_genomic_network}

\item{norm_kos}{Data.table of gene abundances}
}
\value{
Data.table of CMP scores
}
\description{
Calculate CMP scores based on community network and gene abundances
}
\examples{
get_prmt_scores(ko_net[[1]], gene_abunds)
}

