% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleSpec_contributions.R
\name{get_spec_contribs}
\alias{get_spec_contribs}
\title{Perform a series of steps related to identifying single-species contributors.}
\usage{
get_spec_contribs(contrib_file, data_dir, results_file, out_dir, otu_id,
  otu_file, valueVar, make_unnormalized, sum_to_genus, cmps, contributions)
}
\arguments{
\item{contrib_file}{File where metagenome_contributions PICRUSt output is located}

\item{data_dir}{Directory where OTU data is located}

\item{results_file}{File where run_all_metabolites output is located}

\item{out_dir}{Directory where output will be written too}

\item{otu_id}{"All" to evaluate all detected OTUs, otherwise, a vector of OTU IDs to test for contribution}

\item{otu_file}{OTU abundances file}

\item{valueVar}{type of functional abundances to use for single species calculations - either "singleMusicc" or "RelAbundSample"}

\item{make_unnormalized}{The next few parameters are binary variables indicating whether to perform each step of the calculation. This is whether to revert to relative abundance values for the gene abundances from single species}

\item{sum_to_genus}{Whether to sum over OTUs to the genus level (requires Greengenes taxonomy info)}

\item{cmps}{Whether to calculate single-species CMP scores}

\item{contributions}{Whether to evaluate contributions based on single-species CMP scores}

\item{rxn_table}{Community metabolic network template}
}
\value{
No return, saves output to specified file
}
\description{
Perform a series of steps related to identifying single-species contributors.
}
\examples{
read_files(gene_file, met_file)
}

