% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postProcessing.R
\name{get_summary_stats}
\alias{get_summary_stats}
\title{Calculate basic statistics summarizing MIMOSA results}
\usage{
get_summary_stats(node_data, threshold_class = "low")
}
\arguments{
\item{node_data}{Processed table of core MIMOSA results}

\item{threshold_class}{Whether to use the more or less strict significance threshold for counting metabolite outcomes (default less strict)}
}
\value{
List of counts describing 1) total metabolites analyzed, 2) number consistent, 3) percent consistent, 4) number contrasting, 5) percent contrasting
get_summary_stats(node_data)
}
\description{
Calculate basic statistics summarizing MIMOSA results
}
