% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weights.R
\name{get_weights}
\alias{get_weights}
\title{Weights for IMCo model}
\usage{
get_weights(offsets, voxelDims, sigma)
}
\arguments{
\item{offsets}{matrix of offsets from center voxel that define neighborhood}

\item{voxelDims}{voxel dimensions in mm}

\item{sigma}{standard deviation of Gaussian kernel}
}
\value{
vector of weights
}
\description{
Gets weights for IMCo model
}
\examples{
offsets = c(-1, 0, 1)
offsets = expand.grid(offsets, offsets, offsets)
sigma = 3
voxelDims = c(0.25, 0.8, 3)
wts = get_weights(offsets = offsets, voxelDims = voxelDims
sigma = sigma)
stopifnot(length(wts)==27)
}
