% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriorRR.R
\name{getPosteriorResponseRate}
\alias{getPosteriorResponseRate}
\title{Compute the posterior response rate from MCMC samples}
\usage{
getPosteriorResponseRate(x, variable)
}
\arguments{
\item{x}{A MIMOSAResultList or MIMOSAResult. All models should be of type MCMCResult, or fitted using method="mcmc" in MIMOSA.}

\item{variable}{an unquoted variable name in the pData() table of all the models that specifies the grouping variable for which to compute response rates.}
}
\value{
a tibble with the grouping variable, and the 2.5th, median, and 97.5th percentiles of the posterior response rate.
}
\description{
Uses the sampled response indicator when the model class is MCMCResult (or MIMOSAResult encapsulating an MCMCResult) to
compute the posterior response rate distribution. This is summarized to its median, 2.5th and 97.5th credible interval.
}
\details{
The posterior response rate is the correct way to compare response rates across studies and treatment groups, as it is unbiased compared to the response rate computed from hard thresholds
of posterior response probabilities. The credible intervals have the correct behaviour as sample size increases. 

Future versions will allow passing of quantiles for summarization, and maybe the full distribution, depending on needs.
}
