% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helper_functions.R
\name{humann2_format_contributions}
\alias{humann2_format_contributions}
\title{Function for reformatting a Humann2 stratified KO abundance table into a picrust-format table}
\usage{
humann2_format_contributions(path_to_humann_file)
}
\arguments{
\item{path_to_humann_file}{File path to humann2 stratified KO abundance table.}
}
\value{
data.table formatted in the style of a PICRUSt species contribution table compatible with MIMOSA
}
\description{
Function for reformatting a Humann2 stratified KO abundance table into a picrust-format table
}
\examples{
kos_from_species(bv_qpcr, genome_content)
}
