% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ICS}
\alias{ICS}
\title{Stimulated and unstimulated T-cell counts for an ICS assay}
\format{A data frame with 3960 rows}
\description{
A data set containing T-cell counts for various stimulations and cytokines in an ICS assay.
}
\details{
\itemize{
\item pos. The positive cell counts
\item neg. The negative cell counts
\item fname. The feature name (cytokine) measured
\item parent. The parent T-cell population
\item antigen. The antigen stimulation for this sample
\item ID. The subject ID
}
}
\keyword{datasets}

