% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imco.R
\name{imco}
\alias{imco}
\title{Inter-modal Coupling Model}
\usage{
imco(files, brainMask, subMask = NULL, type = "pca", ref = 1,
  neighborhoodSize = 3, verbose = TRUE, retimg = FALSE, outDir = NULL)
}
\arguments{
\item{files}{Vector of full paths to images}

\item{brainMask}{Full path to brain mask}

\item{subMask}{NULL or full path subset of brain mask where coupling should be computed}

\item{type}{"regression" or "pca"}

\item{ref}{reference modality or dependent modality when type="regression"}

\item{neighborhoodSize}{Full width in voxels for FWHM}

\item{verbose}{TRUE for progress bar}

\item{retimg}{If TRUE, return list of estimated coupling maps}

\item{outDir}{Full path to directory where maps should be written}
}
\value{
Estimated IMCo coupling maps, either written to files and/or returned as antsImage objects
}
\description{
Implements full volumetric IMCo coupling estimation on a single subject
}
\examples{
\dontrun{

}
}

