% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imco_reg.R
\name{imco_reg}
\alias{imco_reg}
\title{Regression Inter-modal Coupling Model}
\usage{
imco_reg(files, nhoods, nWts, mask_indices, ref = 1, reverse = FALSE,
  verbose = TRUE, retimg = FALSE, outDir = NULL, propMiss = NULL)
}
\arguments{
\item{files}{Vector of full paths to images or a list of local nifti objects}

\item{nhoods}{Neighborhood matrices computed in imco}

\item{nWts}{Vector of kernel weights}

\item{mask_indices}{Voxels where IMCo is computed}

\item{ref}{Which modality is dependent variable}

\item{reverse}{Should regression be done/returned both Y on X and X on Y? Ignored if # modalities > 2}

\item{verbose}{TRUE for updates on computation, else FALSE}

\item{retimg}{If TRUE, return list of estimated coupling maps as nifti objects}

\item{outDir}{Full path to directory where maps should be written}

\item{propMiss}{Maximum proportion of missing voxels in a neighborhood to tolerate, i.e., return NA if missing more than propMiss in the neighborhood of the center voxel}
}
\value{
Estimated IMCo coupling maps, either written to files and/or returned as nifti objects
}
\description{
Implements full volumetric IMCo coupling estimation on a single subject using regression approach
}
\examples{
\dontrun{

}
}
