% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imco_single.R
\name{imco_single}
\alias{imco_single}
\title{Single Inter-modal Coupling model}
\usage{
imco_single(files, vxl, brainMask, ref = 1, neighborhoodSize = 3)
}
\arguments{
\item{files}{Vector of full paths to images}

\item{vxl}{Vector of single voxel indicies to get data from}

\item{brainMask}{Full path to brain mask}

\item{ref}{reference modality or dependent modality when type="regression"}

\item{neighborhoodSize}{Full width in voxels for FWHM}
}
\value{
Data frame
}
\description{
Data for coupling estimation on a single subject and single voxel
}
\examples{
\dontrun{

}
}

