% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{kos_from_species}
\alias{kos_from_species}
\title{Function for getting ko abundances from BV qPCR data}
\usage{
kos_from_species(qpcr, ref_kos, scale_factor = 1)
}
\arguments{
\item{qpcr}{qPCR species abundance data across samples}

\item{ref_kos}{gene content of every species}

\item{scale_factor}{Scale abundances by a factor}
}
\value{
gene abundances across samples
}
\description{
Function for getting ko abundances from BV qPCR data
}
\examples{
kos_from_species(bv_qpcr, genome_content)
}

