% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ants_image.R
\name{make_ants_image}
\alias{make_ants_image}
\title{make_ants_image}
\usage{
make_ants_image(vec, mask_indices, reference)
}
\arguments{
\item{vec}{vector of IMCo values}

\item{mask_indices}{voxels where IMCo was computed}

\item{reference}{image for header info}
}
\value{
3D ANTS image of IMCo measurements
}
\description{
Convert vector of IMCo into ANTS image
}
\examples{
\dontrun{

}
}
