% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helper_functions.R
\name{make_contrib_network}
\alias{make_contrib_network}
\title{Convert edge list and node list to class network object for plotting with ggnetwork}
\usage{
make_contrib_network(gene_contribs, node_data, node_attrs, edge_attrs)
}
\arguments{
\item{gene_contribs}{output of gene_contributions}

\item{node_data}{output of run_all_metabolites}

\item{node_attrs}{vector of node attributes to include in network}

\item{edge_attrs}{vector of edge attributes to include in network}
}
\value{
network object
}
\description{
Convert edge list and node list to class network object for plotting with ggnetwork
}
\examples{
make_contrib_network(gene_contribs_all, node_data, c("Dataset", "Correlation", "Metabolite", "PredictionType2", "SuperPath"), c("KO", "Cor", "Dataset", "stoichReac", "stoichProd"))
}
