% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postProcessing.R
\name{make_gene_species_contributor_table}
\alias{make_gene_species_contributor_table}
\title{Make table of species-specific gene and reaction contributors identified by MIMOSA}
\usage{
make_gene_species_contributor_table(geneContribs, species_ref, node_data_spec)
}
\arguments{
\item{geneContribs}{Table from MIMOSA gene/rxn contributor analysis ("_geneContribAnalysis.txt")}

\item{species_ref}{Table of genome content as produced by PICRUSt}

\item{node_data_spec}{Table of core metabolite results and their species contributors}
}
\value{
Table of species-specific reaction contributors and overall MIMOSA results for each analyzed metabolite
make_gene_species_contributor_table(gene_contribs, species_ref, node_data_species)
}
\description{
Make table of species-specific gene and reaction contributors identified by MIMOSA
}
