% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle_network.R
\name{make_network_matrix}
\alias{make_network_matrix}
\title{Convert network edge list to stoichiometric matrix}
\usage{
make_network_matrix(rxn_table)
}
\arguments{
\item{rxn_table}{edge list network}
}
\value{
list of two matrices of compounds x genes, one with NAs to indicate lack of relationship vs. neutral and one with all zeros
}
\description{
Convert network edge list to stoichiometric matrix
}
\examples{
make_network_matrix(ko_net[[3]])
}

