% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{make_pairwise_met_matrix}
\alias{make_pairwise_met_matrix}
\title{Make a data vector into a pairwise difference matrix (can be used for either PRMT scores or metabolite concentrations)}
\usage{
make_pairwise_met_matrix(metabolite, met_mat, diff_function = "difference")
}
\arguments{
\item{metabolite}{metabolite ID}

\item{met_mat}{Abundances or scores for that metabolite across samples}

\item{diff_function}{Difference by default, could be fold_change}
}
\value{
Data.frame of score or abundance pairwise differentials
}
\description{
Make a data vector into a pairwise difference matrix (can be used for either PRMT scores or metabolite concentrations)
}
\examples{

}

