% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{make_perm_mat}
\alias{make_perm_mat}
\title{Make matrix of permutations for mantel test with a row for each permutation and a column for each sample}
\usage{
make_perm_mat(nSamples, nPerm)
}
\arguments{
\item{nSamples}{number of samples}

\item{nPerm}{number of permutations}
}
\description{
Make matrix of permutations for mantel test with a row for each permutation and a column for each sample
}

