% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helper_functions.R
\name{mantel_2sided}
\alias{mantel_2sided}
\title{Modification of vegan mantel test to test 2-sided or significantly less than}
\usage{
mantel_2sided(xdis, ydis, method = "pearson", permutations = 999,
  strata, na.rm = FALSE, direction = "pos")
}
\arguments{
\item{xdis}{A distance matrix}

\item{ydis}{Another distance matrix}

\item{method}{Correlation coefficient to use (pearson or spearman)}

\item{permutations}{Number of permutations to perform}
}
\value{
Data.frame of score or abundance pairwise differentials
}
\description{
Modification of vegan mantel test to test 2-sided or significantly less than
}
\examples{

}
