\docType{class}
\name{MCMCResult}
\alias{MCMCResult}
\title{The output of fitting the Dirichlet--Multinomial MIMOSA model via MCMC}
\usage{
  MCMCResult(object = NULL)
}
\arguments{
  \item{object}{of class
  \code{MDMixResult},\code{MCMCResult},or
  \code{BetaMixResult}}
}
\description{
  Fitting the model using the \code{MIMOSA} function
  returns an object of this class.

  This constructor is used by the \code{MIMOSA} model
  fitting method.
}

