% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helper_functions.R
\name{met_names}
\alias{met_names}
\title{Get metabolite name for KEGG ID(s)}
\usage{
met_names(met_ids)
}
\arguments{
\item{met_ids}{One or a vector of KEGG compound IDs}
}
\value{
vector of character metabolite names
}
\description{
Get metabolite name for KEGG ID(s)
}
\examples{
met_names(c("C00082", "C00334"))

}

