\name{MIMOSA-package}
\alias{MIMOSA-package}
\alias{MIMOSA}
\docType{package}
\title{
Fits a Bayesian mixture model to identify differences in cell population counts between treatment groups for data typically generated by flow cytometry intracellular cytokine staining assays.
}
\description{
Fits a mixture of beta--binomials model to two--dimensional contingency table data typically generated for Intracellular Cytokine Staining (ICS) and other flow cytometry assays. These are typically termed "positivity calls". Could be used for any two dimensional contingency table data.
}
\details{
Package:  MIMOSA
Type:  Package
Version:  0.1.0
Date:  2011-10-26
License:  Artistic 2.0
LazyLoad:  yes

This package fits a mixture of beta-binomial models to contingency table data typically generated from Intracellular Cytokine Staining (ICS) flow cytometry assays uses to measure immune response to vaccines or other immunological challenges. However, it can be used with any type of two--dimensional contingency table data where there are multiple observations.

The model uses a Bayesian approach to borrow strength across individuals. In the ICS setting, typically, positive and negative cell events are counted for a stimulated and an unstimulated (i.e. treated and untreated) sample from a single individual. The goal of the assay is to determine if (antigen) stimulation leads to an increase in positive cytokine staining cells when compared to the unstimulated sample. If the proportion of positive cells is greater for the stimulation, this oberservation / individual is termed a responder, and the process of identifying responders are termed "positivity calls".

This package models the positive stimulated and unstimulated counts as draws from two binomial distributions with proportions ps and pu. Beta priors are placed on the ps and pu with hyperparameters alpha0, beta0, alphaS, and betaS. The observed data are a mixture of two generative processes, either ps=pu, and thus alpha0=alphaS, beta0=betaS, or ps>pu and thus alphaS!=alpha0 and betaS!=beta0. The resulting mixture model is fit via EM and the parameter estimates for the model are returned. Posterior estimates of ps and pu for any observation (individual) can be estimated via Gibbs sampling. The unified Bayesian framework borrows strength across observations, improving the positivity calls. 

The main function for model fitting is \code{BetaMix()}. It fits the model via the EM algorithm given the positive and negative counts for stimulated and unstimulated samples, and multiple observations.
}
\author{
Greg Finak <gfinak@fhcrc.org>

Maintainer: Greg Finak <gfinak@fhcrc.org>

}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ models }
\keyword{ design }
\seealso{
\code{\link{BetaMix}}
}
\examples{
#TODO
}
