% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mimosa_fit.R
\name{mimosa_fit}
\alias{mimosa_fit}
\title{MIMoSA Fit on Training Data}
\usage{
mimosa_fit(training_dataframe, formula)
}
\arguments{
\item{training_dataframe}{data.frame(s) produced by the mimosa_data function}

\item{formula}{formula to be fit by glm model}
}
\value{
Returns a glm object containing the trained MIMoSA coefficients.
}
\description{
This function trains the MIMoSA model from a data.frame produced by an element from the output of the function mimosa_data.
}
