% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mimosa_formula.R
\name{mimosa_formula}
\alias{mimosa_formula}
\title{MIMoSA Model Formula}
\usage{
mimosa_formula(T2 = TRUE, PD = TRUE)
}
\arguments{
\item{T2}{Logical. Should the terms including \code{T2} be included?}

\item{PD}{Logical. Should the terms including \code{PD} be included?}
}
\value{
A \code{formula} object
}
\description{
MIMoSA Model Formula
}
\examples{
mimosa_formula()
mimosa_formula(T2 = FALSE)
mimosa_formula(PD = FALSE)
mimosa_formula(T2 = FALSE, PD = FALSE)
}
