% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mimosa_model.R
\docType{data}
\name{mimosa_model}
\alias{mimosa_model}
\title{MIMoSA trained model when all images are input}
\usage{
mimosa_model
}
\description{
Trained MIMoSA model for when T1, T2, FLAIR, and PD images are available
}
\keyword{datasets,}
\keyword{model}
\keyword{trained,}
