% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mimosa_model_No_PD.R
\docType{data}
\name{mimosa_model_No_PD}
\alias{mimosa_model_No_PD}
\title{MIMoSA trained model when PD sequence is missing}
\usage{
mimosa_model_No_PD
}
\description{
Trained MIMoSA model for when T1, T2, and FLAIR images are available
}
\keyword{datasets,}
\keyword{model}
\keyword{trained,}
