% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mimosa_model_No_PD_T2.R
\docType{data}
\name{mimosa_model_No_PD_T2}
\alias{mimosa_model_No_PD_T2}
\title{MIMoSA trained model when PD and T2 sequences are missing}
\usage{
mimosa_model_No_PD_T2
}
\description{
Trained MIMoSA model for when T1 and FLAIR images are available
}
\keyword{datasets,}
\keyword{model}
\keyword{trained,}
