% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mimosa_model_No_T2.R
\docType{data}
\name{mimosa_model_No_T2}
\alias{mimosa_model_No_T2}
\title{MIMoSA trained model when T2 sequence is missing}
\usage{
mimosa_model_No_T2
}
\description{
Trained MIMoSA model for when T1, FLAIR, and PD images are available
}
\keyword{datasets,}
\keyword{model}
\keyword{trained,}
