% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIMOSA.R
\name{MIMOSAExpressionSet}
\alias{MIMOSAExpressionSet}
\title{Construct an ExpressionSet for MIMOSA}
\usage{
MIMOSAExpressionSet(df, featureCols)
}
\arguments{
\item{df}{a data.frame that is in the correct form}

\item{featureCols}{the indices of the columns that identify features.}
}
\description{
Starting from a reshaped data frame in the correct format, construct
an ExpressionSet object that can be used with MIMOSA.
}
\details{
The featureCols will be used to construct feature names, and these
 columns will be dropped from the exprs matrix. The column names are assumed
 to have names that contain '_' characters separating phenotypic
 characteristics. These would be generated automatically
 if the data frame was constrcuted with 'reshape'. They
 are used to construct the phenoData for the expression set
}
\examples{
E<-ConstructMIMOSAExpressionSet(ICS,
  reference=ANTIGEN\%in\%'negctrl',measure.columns=c('CYTNUM','NSUB'),
  other.annotations=c('CYTOKINE','TCELLSUBSET','ANTIGEN','UID'),
  default.cast.formula=component~UID+ANTIGEN+CYTOKINE+TCELLSUBSET,
  .variables=.(TCELLSUBSET,CYTOKINE,UID),
  featureCols=1,ref.append.replace='_REF')
}
