% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{MIMOSAResult-class}
\alias{MIMOSAResult-class}
\title{Stores the result of a MIMOSA fitted model}
\description{
\code{MIMOSA} returns an object of \code{MIMOSAResult} irrespective of which
method / implementation is used to fit the data.
}

