\docType{class}
\name{MIMOSAResult}
\alias{MIMOSAResult}
\alias{show}
\alias{show,MIMOSAResult-method}
\title{Stores the result of a MIMOSA fitted model}
\description{
  \code{MIMOSA} returns an object of \code{MIMOSAResult}
  irrespective of which method / implementation is used to
  fit the data.
}

