% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIMOSA.R
\docType{methods}
\name{pData<-,MCMCResult,ANY-method}
\alias{pData<-,MCMCResult,ANY-method}
\alias{pData<-,MIMOSAResult,ANY-method}
\alias{pData<-,BetaMixResult,data.frame-method}
\alias{pData,MIMOSAResult-method}
\alias{pData,MDMixResult-method}
\alias{pData,BetaMixResult}
\alias{pData,MCMCResult-method}
\alias{pData,BetaMixResult-method}
\alias{pData.MIMOSAResultList}
\alias{pData,MIMOSAResultList-method}
\title{pData}
\usage{
\S4method{pData}{MCMCResult,ANY}(object) <- value

\S4method{pData}{MIMOSAResult,ANY}(object) <- value

\S4method{pData}{BetaMixResult,data.frame}(object) <- value

\S4method{pData}{MIMOSAResult}(object)

\S4method{pData}{MDMixResult}(object)

\S4method{pData}{MCMCResult}(object)

\S4method{pData}{BetaMixResult}(object)

\method{pData}{MIMOSAResultList}(object)

\S4method{pData}{MIMOSAResultList}(object)
}
\arguments{
\item{object}{is the MIMOSAResult returned from a call to MIMOSA}

\item{value}{the phenoData table to be assigned to the object.}
}
\value{
an object of type \code{data.frame}
}
\description{
pData

pData extract the phenoData table from a MIMOSA result
}
\details{
Extracts the phenoData data.frame from a MIMOSAResult object
}
