% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{pData,MIMOSAResult-method}
\alias{pData,MCMCResult-method}
\alias{pData,MDMixResult-method}
\alias{pData,MIMOSAResult-method}
\alias{pData,MIMOSAResultList-method}
\alias{pData.MIMOSAResultList}
\title{pData extract the phenoData table from a MIMOSA result}
\usage{
\S4method{pData}{MIMOSAResult}(object)

\S4method{pData}{MDMixResult}(object)

\S4method{pData}{MCMCResult}(object)

pData.MIMOSAResultList(object)

\S4method{pData}{MIMOSAResultList}(object)
}
\arguments{
\item{object}{is the MIMOSAResult returned from a call to MIMOSA}
}
\value{
an object of type \code{data.frame}
}
\description{
pData extract the phenoData table from a MIMOSA result
}
\details{
Extracts the phenoData data.frame from a MIMOSAResult object
}

