% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postProcessing.R
\name{plot_contrib_net}
\alias{plot_contrib_net}
\title{Use ggnetwork to make a network visualization of metabolite outcomes and primary gene/reaction contributors}
\usage{
plot_contrib_net(net_obj, col_attr, node_size_attr, edge_size_attr, node_lab,
  edge_lab)
}
\arguments{
\item{gene_contribs}{output of gene_contributions}

\item{node_data}{output of run_all_metabolites}

\item{node_attrs}{vector of node attributes to include in network}

\item{edge_attrs}{vector of edge attributes to include in network}
}
\value{
network object
}
\description{
Use ggnetwork to make a network visualization of metabolite outcomes and primary gene/reaction contributors
}
\examples{
plot_contrib_net(gene_contribs_all, node_data, c("Dataset", "Correlation", "Metabolite", "PredictionType2", "SuperPath"), c("KO", "Cor", "Dataset", "stoichReac", "stoichProd"))
}
