% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postProcessing.R
\name{plot_metabolite_concentrations}
\alias{plot_metabolite_concentrations}
\title{Make heatmap plot of metabolite concentrations across samples for analyzed compounds}
\usage{
plot_metabolite_concentrations(mets, met_list)
}
\arguments{
\item{mets}{Table of metabolite abundances}

\item{met_list}{List of metabolite IDs to include in the plot}
}
\value{
plot of metabolite concentrations across samples
plot_metabolite_concentrations(mets)
}
\description{
Make heatmap plot of metabolite concentrations across samples for analyzed compounds
}
