% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{popular2}
\alias{popular2}
\title{Simulated popularity data from Hox et al. (2018)}
\format{
## `popular2`
A data frame with 2k rows and 15 columns:
\describe{
  \item{pupil}{}
  \item{class}{}
  \item{extrav}{}
  \item{sex}{}
  \item{texp}{}
  \item{popular}{}
  \item{popteach}{}
  \item{Zextrav}{}
  \item{Zsex}{}
  \item{Ztexp}{}
  \item{Zpopular}{}
  \item{Zpopteach}{}
  \item{Cextrav}{}
  \item{Ctexp}{}
  \item{Csex}{}
}
}
\source{
<https://github.com/MultiLevelAnalysis/Datasets-third-edition-Multilevel-book/tree/master/chapter%202/popularity>
}
\usage{
popular2
}
\description{
The popularity data in popular2 are simulated data for 2000 pupils in 100
schools. The purpose is to offer a very simple example for multilevel
regression analysis. The main outcome variable is the pupil popularity, a
popularity rating on a scale of 1–10 derived by a sociometric procedure.
Typically, a sociometric procedure asks all pupils in a class to rate all the
other pupils, and then assigns the average received popularity rating to each
pupil. Because of the sociometric procedure, group effects as apparent from
higher-level variance components are rather strong. There is a second outcome
variable: pupil popularity as rated by their teacher, on a scale from 1 to
10. The explanatory variables are pupil gender (boy = 0, girl = 1), pupil
extraversion (10-point scale), and teacher experience in years.
}
\keyword{datasets}
