% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postProcessing.R
\name{print_taxa_contributor_table}
\alias{print_taxa_contributor_table}
\title{Return summarized and concise data on taxonomic contributors for consistent metabolites}
\usage{
print_taxa_contributor_table(node_data_spec)
}
\arguments{
\item{node_data_spec}{Table of species contributors and core MIMOSA results}
}
\value{
Table of summarized and rounded data on taxonomic contributors for consistent metabolites
}
\description{
Return summarized and concise data on taxonomic contributors for consistent metabolites
}
