% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleSpec_contributions.R
\name{prmt_species_contributions}
\alias{prmt_species_contributions}
\title{Evaluate species contributors for a single metabolite with qPCR/species abundance data}
\usage{
prmt_species_contributions(j, prmts_sub_good, all_rxns, subjects, norm_kos,
  ko_net, qpcr, ref_kos, cor_with = F)
}
\arguments{
\item{j}{metabolite # (usually from lapply/sapply)}

\item{prmts_sub_good}{CMP scores for metabolites with abundance data}

\item{all_rxns}{list of relevant reactions for each metabolite}

\item{subjects}{vector of subjects}

\item{norm_kos}{data.table of gene abundances}

\item{ko_net}{network created by generate_genomic_network}

\item{qpcr}{original species abundances}

\item{ref_kos}{gene abundances for each species}

\item{cor_with}{whether to look at the correlation of CMP scores of each species by itself with the metabolite, or of the whole community with that species removed}
}
\value{
list of potential key species contributors and associated info
}
\description{
Evaluate species contributors for a single metabolite with qPCR/species abundance data
}
\examples{
sapply(1:length(metabolites), prmt_species_contributions, cmp_scores, all_rxns,
all_subjects, ko_abunds, ko_net, spec_abunds, ref_kos)
}

