% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postProcessing.R
\name{process_gene_contribs}
\alias{process_gene_contribs}
\title{Process table of gene and reaction contributors}
\usage{
process_gene_contribs(geneContribs, node_data)
}
\arguments{
\item{geneContribs}{Table from MIMOSA gene/rxn contributor analysis ("_geneContribAnalysis.txt")}

\item{node_data}{Table of core metabolite results}
}
\value{
Merged table of MIMOSA results and gene and reaction contributors
process_gene_contribs(gene_contribs, node_data)
}
\description{
Process table of gene and reaction contributors
}
