% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postProcessing.R
\name{process_metabolite_summary}
\alias{process_metabolite_summary}
\title{Add summary columns to metabolite 'node' file of MIMOSA results}
\usage{
process_metabolite_summary(node_data, thresholds, ko_summary)
}
\arguments{
\item{node_data}{Table from "_nodes.txt" file produced by runMimosa.R}

\item{thresholds}{Significance thresholds (q-value and p-value) to be used for classifying metabolites as consistent or contrasting. By default this function classifies based on 2 different thresholds}

\item{ko_summary}{Table of gene contributor information produced by MIMOSA (geneContribCompoundSummary.txt))}

\item{path_key}{Table of metabolite categories}
}
\value{
Processed table of merged results for each metabolite
process_metabolite_summary(node_data, list(c(0.01, 0.01), c(0.1, 0.1)))
}
\description{
Add summary columns to metabolite 'node' file of MIMOSA results
}
