% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle_network.R
\name{randomize_net}
\alias{randomize_net}
\title{Function to randomize a network}
\usage{
randomize_net(netw, n_reps)
}
\arguments{
\item{netw}{edge list network}

\item{n_reps}{number of times to successfully sample and switch}
}
\value{
randomized edge list
}
\description{
Randomly sample 2 edges and if it works, switch products
}
\examples{
randomize_net(allnet1, 1000)
}

