% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{read_files}
\alias{read_files}
\title{Read in processed files and assign keys to datasets. Sets NA metabolite values to 0.}
\usage{
read_files(genefile, metfile)
}
\arguments{
\item{genefile}{File where gene data is located.}

\item{metfile}{File where metaboltie data is located.}
}
\value{
A list in which the first item is a data.table of gene abundances and the second is a data.table of metabolite abundances.
}
\description{
Read in processed files and assign keys to datasets. Sets NA metabolite values to 0.
}
\examples{
read_files(gene_file, met_file)
}

