\name{repmat}
\alias{repmat}
\title{R equivalent of repmat}
\usage{
  repmat(X, m, n)
}
\arguments{
  \item{X}{input matrix}

  \item{m}{number times to replicate the rows}

  \item{n}{number of times to replicate the columns}
}
\description{
  R equivalent of repmat
}
\details{
  replicates mx rows x m and nx columns by n
}

