% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Starts the mimosa shiny app}
\usage{
run_app(host = getOption("shiny.host"), port = getOption("shiny.port"))
}
\arguments{
\item{host}{default is from getOption shiny.host}

\item{port}{default is from getOption shiny.port}
}
\description{
Starts the mimosa shiny app
}
