% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{run_shuffle}
\alias{run_shuffle}
\title{Run one iteration of MIMOSA analysis with a randomized metabolic network. Run this function many times to generate a null distribution of metabolite comparisons.}
\usage{
run_shuffle(out_file, id_num = 1, n_iter = 5000, nonzero_filter = 3,
  qval_thresholds = c(0.1, 0.01))
}
\arguments{
\item{out_file}{File path to .Rda file of MIMOSA output from run_all_metabolites}

\item{id_num}{ID number of shuffle (if running many iterations)}

\item{n_iter}{Number of edge switches to generate randomized network, default is 5000}

\item{nonzero_filter}{Minimum number of samples required to have nonzero concentrations and nonzero metabolic potential scores in order for metabolite to be evaluated, default is 3}

\item{qval_thresholds}{1 or more significance thresholds above which to count the number of metabolites}
}
\value{
None, writes two tables to file - one of the total counts of metabolites meeting each threshold, and one of the results for each metabolite
}
\description{
Run one iteration of MIMOSA analysis with a randomized metabolic network. Run this function many times to generate a null distribution of metabolite comparisons.
}
\examples{
run_shuffle("Dataset2_bv_out.rda", id_num = 1)
}
