% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleSpec_contributions.R
\name{save_contribs_by_species}
\alias{save_contribs_by_species}
\title{Get contributors for every metabolite and save}
\usage{
save_contribs_by_species(contribs, valueVar, out_dir)
}
\arguments{
\item{contribs}{data.table of OTU, genes, samples and abundances}

\item{valueVar}{"relAbundSample" or "singleMusicc", abundance metric to use for single taxon gene abundances}

\item{out_dir}{directory to write to}
}
\value{
Null, saves to Rdata file
}
\description{
Get contributors for every metabolite and save
}

