% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIMOSA.R
\name{show}
\alias{show}
\alias{show,MIMOSAResult-method}
\title{show}
\usage{
\S4method{show}{MIMOSAResult}(object)
}
\arguments{
\item{object}{\code{MIMOSAResultList}}
}
\description{
Show a MIMOSAResultList
}
\details{
Show a summary of a MIMOSAResultList.
}
