% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{single_gene_cmp}
\alias{single_gene_cmp}
\title{Calculate scores based on only a single gene.}
\usage{
single_gene_cmp(compound, gene, norm_kos, ko_net)
}
\description{
Calculate scores based on only a single gene.
}
