% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleSpec_contributions.R
\name{single_spec_musicc}
\alias{single_spec_musicc}
\title{Ensure contribution table abundances are relative otu abundance * copy number}
\usage{
single_spec_musicc(contribs)
}
\arguments{
\item{contribs}{data.table of OTU, genes, samples and abundances}
}
\value{
contribs with abundances normalized to copy number
}
\description{
Ensure contribution table abundances are relative otu abundance * copy number
}
