% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleSpec_contributions.R
\name{sum_to_genus}
\alias{sum_to_genus}
\title{Sum OTU data to the genus level}
\usage{
sum_to_genus(contribs, valueVar, taxonomy)
}
\arguments{
\item{contribs}{metagenome_contributions output of PICRUSt}

\item{valueVar}{"relAbundSample" or "singleMusicc", abundance metric to use for single taxon gene abundances}
}
\value{
contribs added to the genus level
}
\description{
Needs Greengenes 97_otu_taxonomy.txt to be in the same directory
}
