% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postProcessing.R
\name{taxonomic_contributor_heatmap_plot_grid}
\alias{taxonomic_contributor_heatmap_plot_grid}
\title{Make plot of individual metabolites and their taxonomic contributors}
\usage{
taxonomic_contributor_heatmap_plot_grid(node_data, node_data_spec,
  prediction_colors = c("#1B9E77", "lightgrey", "#F46D43"))
}
\arguments{
\item{node_data}{Table of core MIMOSA results}

\item{node_data_spec}{Table of MIMOSA taxonomic contributors}

\item{prediction_colors}{List of 3 colors for generating overall correlation color scale}
}
\value{
Plot object showing overall MIMOSA correlation and individual taxonomic contributors for each metabolite
}
\description{
Make plot of individual metabolites and their taxonomic contributors
}
