% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{test_met_enrichment}
\alias{test_met_enrichment}
\title{Test against list of known microbial metabolites}
\usage{
test_met_enrichment(node_data, met_list)
}
\arguments{
\item{node_data}{output of run_all_metabolites}

\item{met_list}{List of metabolites to test against}
}
\value{
Null, just prints a bunch of p-values
}
\description{
Test against list of known microbial metabolites
}
\examples{
test_met_enrichment(node_data, microbial_mets)
}

