% Generated by roxygen2 (4.0.2): do not edit by hand
\name{volcanoPlot}
\alias{volcanoPlot}
\title{Volcano plot for a MIMOSA model}
\usage{
volcanoPlot(x, effect_expression = NA, facet_var = NA, threshold = 0.01)
}
\arguments{
\item{x}{A \code{MIMOSAResultList}}

\item{effect_expression}{an \code{expression} that defines the effect size. Usually a function of the stimulated and unstimulated proportions from \code{countsTable(x,proportion=TRUE)}}

\item{facet_var}{an \code{expression} defining the faceting in ggplot parlance. i.e. \code{~ faceting + variables}}

\item{threshold}{a \code{numeric} value between [0,1] for coloring significant observations (based on the q-value)}
}
\description{
Plots effect size vs posterior probablilty of response from a MIMOSAResultList, faceting by the conditioning variables.
}
\examples{
data(ICS)
E<-ConstructMIMOSAExpressionSet(ICS,
  reference=ANTIGEN\%in\%'negctrl',measure.columns=c('CYTNUM','NSUB'),
  other.annotations=c('CYTOKINE','TCELLSUBSET','ANTIGEN','UID'),
  default.cast.formula=component~UID+ANTIGEN+CYTOKINE+TCELLSUBSET,
  .variables=.(TCELLSUBSET,CYTOKINE,UID),
  featureCols=1,ref.append.replace='_REF')

result<-MIMOSA(NSUB+CYTNUM~UID+TCELLSUBSET+CYTOKINE|ANTIGEN,
    data=E, method='EM',
    subset=RefTreat\%in\%'Treatment'&ANTIGEN\%in\%'ENV',
    ref=ANTIGEN\%in\%'ENV'&RefTreat\%in\%'Reference')
volcanoPlot(result,CYTNUM-CYTNUM_REF)
}
\seealso{
\code{\link{countsTable}}
}

