% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_slr.R
\name{weighted_slr}
\alias{weighted_slr}
\title{Weighted simple linear regression}
\usage{
weighted_slr(x, y, wts, mInds, refImg)
}
\arguments{
\item{x}{dependent modality matrix of nhood values (primary)}

\item{y}{independent modality matrix of nhood values (primary)}

\item{wts}{vector of kernel weights}

\item{mInds}{indices for which to compute IMCo}

\item{refImg}{reference image for header info}
}
\value{
L list including intercept and slope from both weighted SLRs
}
\description{
Returns parameters from weighted regression of x on y (primary) and y on x (secondary)
}
\examples{
\dontrun{

}
}
