% Generated by roxygen2 (4.0.2): do not edit by hand
\name{boxplotMIMOSAResultList}
\alias{boxplotMIMOSAResultList}
\title{boxplotMIMOSAResultList}
\usage{
boxplotMIMOSAResultList(data, title = "A Boxplot", x_axis_category = NULL,
  cofactor = 5000)
}
\arguments{
\item{data}{\code{MIMOSAResultList}}

\item{title}{\code{character} Title of the plot.}

\item{x_axis_category}{\code{name} the column of the phenoData frame for the x-axis of the boxplots.}

\item{cofactor}{\code{integer} cofactor of the arcsinhTransform for the y axis.}
}
\value{
\code{ggplot} object.
}
\description{
Boxplots of MIMOSA
}
\details{
Generate boxplots for MIMOSA positivity calls.
}
\author{
Greg Finak
}

