% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{find_id}
\alias{find_id}
\title{Tries to find the grouping variable (id)}
\usage{
find_id(d)
}
\arguments{
\item{d}{data frame}
}
\value{
potential grouping variables
}
\description{
Works heuristically by (1) checking how many level-2-variables are present
for every potential grouping variable and (2) how many reverse levels are
present for every potential grouping variable.
}
\examples{
find_id(mlmRev::Exam)
}
