% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{gene_contributions}
\alias{gene_contributions}
\title{Identify potential important gene contributors for each metabolite.}
\usage{
gene_contributions(j, prmts_sub_good, all_rxns, subjects, norm_kos, ko_net)
}
\value{
list of contributors and correlations
}
\description{
Identify potential important gene contributors for each metabolite.
}
\examples{

}

