% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helper_functions.R
\name{get_non_rev_rxns}
\alias{get_non_rev_rxns}
\title{Get non-reverible reactions of network.}
\usage{
get_non_rev_rxns(rxn_table, all_rxns = T)
}
\arguments{
\item{pvals}{Vector of p-values}

\item{method}{Must be either "fdr" or "bonferroni"}
}
\value{
Vector of corrected values
}
\description{
Get non-reverible reactions of network.
}
\examples{
get_non_rev_rxns(rxn_table)
}
