% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleSpec_contributions.R
\name{get_spec_contribs}
\alias{get_spec_contribs}
\title{Perform a series of steps related to identifying single-species contributors.}
\usage{
get_spec_contribs(contrib_file, data_dir, results_file, out_dir,
  otu_id = "all", otu_file, valueVar = "singleMusicc", sum_to_genus,
  write_out = T)
}
\arguments{
\item{contrib_file}{File where metagenome_contributions PICRUSt output is located}

\item{data_dir}{Directory where OTU data is located}

\item{results_file}{File where run_all_metabolites output is located}

\item{out_dir}{Directory where output will be written too}

\item{otu_id}{"All" to evaluate all detected OTUs, otherwise, a vector of OTU IDs to test for contribution}

\item{otu_file}{OTU abundances file}

\item{valueVar}{type of functional abundances to use for single species calculations - either "singleMusicc" or "RelAbundSample"}

\item{sum_to_genus}{Whether to sum over OTUs to the genus level (requires Greengenes taxonomy info)}
}
\value{
table of species, metabolites, whether that species is a contributor for that metabolite and the correlation strength
}
\description{
Perform a series of steps related to identifying single-species contributors.
}
\examples{
read_files(gene_file, met_file)
}

