% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imco.R
\name{imco}
\alias{imco}
\title{Inter-modal Coupling Model}
\usage{
imco(files, brainMask, subMask = NULL, type = "pca", ref = 1,
  neighborhoodSize = 3, reverse = TRUE, verbose = TRUE, retimg = FALSE,
  outDir = NULL)
}
\arguments{
\item{files}{Vector of full paths to images or a list of local nifti objects}

\item{brainMask}{Full path to brain mask or local nifti object}

\item{subMask}{NULL or local nifti object or the full path to a mask that is a subset of brain mask where coupling should be computed}

\item{type}{"regression" or "pca"}

\item{ref}{reference modality when type="pca" or dependent modality when type="regression"}

\item{neighborhoodSize}{Full width in voxels for FWHM}

\item{reverse}{If TRUE, calculates both regressions if type="regression" and length(files)==2, otherwise ignored}

\item{verbose}{TRUE for updates on computation, else FALSE}

\item{retimg}{If TRUE, return list of estimated coupling maps as nifti objects}

\item{outDir}{Full path to directory where maps should be written}
}
\value{
Estimated IMCo coupling maps, either written to files and/or returned as nifti objects
}
\description{
Implements full volumetric IMCo coupling estimation on a single subject
}
\examples{
\dontrun{

}
}
