% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imco_single.R
\name{imco_single}
\alias{imco_single}
\title{Single Inter-modal Coupling model}
\usage{
imco_single(files, vxl, brainMask, subMask = NULL, ref = 1,
  neighborhoodSize = 3)
}
\arguments{
\item{files}{Vector of full paths to images or a list of local nifti objects}

\item{vxl}{Vector of single voxel indices to get data from}

\item{brainMask}{Full path to brain mask or local nifti object}

\item{subMask}{NULL or local nifti object or the full path to a mask that is a subset of brain mask where coupling should be computed}

\item{ref}{reference modality when type="pca" or dependent modality when type="regression"}

\item{neighborhoodSize}{Full width in voxels for FWHM}
}
\value{
Data frame
}
\description{
Data for coupling estimation on a single subject and single voxel
}
