% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helper_functions.R
\name{kos_from_species}
\alias{kos_from_species}
\title{Function for getting ko abundances from species abundance data and table of reference genomes}
\usage{
kos_from_species(spec_abunds, ref_kos, scale_factor = 1)
}
\arguments{
\item{spec_abunds}{Species abundance data table across samples}

\item{ref_kos}{gene content of every species}

\item{scale_factor}{Scale abundances by a factor}
}
\value{
gene abundances across samples
}
\description{
Function for getting ko abundances from species abundance data and table of reference genomes
}
\examples{
kos_from_species(bv_qpcr, genome_content)
}

