% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getZ}
\alias{getW}
\alias{getW.MIMOSAResult}
\alias{getW.MIMOSAResultList}
\alias{getZ}
\alias{getZ.MIMOSAResult}
\alias{getZ.MIMOSAResultList}
\title{Extract the posterior probabilities of response from a MIMOSA model}
\usage{
getZ(x)

\method{getZ}{MIMOSAResultList}(x)

\method{getZ}{MIMOSAResult}(x)

getW(x)

\method{getW}{MIMOSAResultList}(x)

\method{getW}{MIMOSAResult}(x)
}
\arguments{
\item{x}{output from a MIMOSA model}
}
\value{
a \code{matrix} of posterior probabilities

a \code{vector} of component weights
}
\description{
Extract the posterior probabilities of response from a MIMOSA model

Extract the component weights from a MIMOSA model
}
\examples{
data(ICS)
E<-ConstructMIMOSAExpressionSet(ICS,
  reference=ANTIGEN\%in\%'negctrl',measure.columns=c('CYTNUM','NSUB'),
  other.annotations=c('CYTOKINE','TCELLSUBSET','ANTIGEN','UID'),
  default.cast.formula=component~UID+ANTIGEN+CYTOKINE+TCELLSUBSET,
  .variables=.(TCELLSUBSET,CYTOKINE,UID),
  featureCols=1,ref.append.replace='_REF')

result<-MIMOSA(NSUB+CYTNUM~UID+TCELLSUBSET+CYTOKINE|ANTIGEN,
    data=E, method='EM',
    subset=RefTreat\%in\%'Treatment'&ANTIGEN\%in\%'ENV',
    ref=ANTIGEN\%in\%'ENV'&RefTreat\%in\%'Reference')
    getZ(result)
data(ICS)
E<-ConstructMIMOSAExpressionSet(ICS,
  reference=ANTIGEN\%in\%'negctrl',measure.columns=c('CYTNUM','NSUB'),
  other.annotations=c('CYTOKINE','TCELLSUBSET','ANTIGEN','UID'),
  default.cast.formula=component~UID+ANTIGEN+CYTOKINE+TCELLSUBSET,
  .variables=.(TCELLSUBSET,CYTOKINE,UID),
  featureCols=1,ref.append.replace='_REF')

result<-MIMOSA(NSUB+CYTNUM~UID+TCELLSUBSET+CYTOKINE|ANTIGEN,
    data=E, method='EM',
    subset=RefTreat\%in\%'Treatment'&ANTIGEN\%in\%'ENV',
    ref=ANTIGEN\%in\%'ENV'&RefTreat\%in\%'Reference')
getW(result)
}

