\docType{package}
\name{MIMOSA-package}
\alias{MIMOSA-package}
\title{MIMOSA: Mixture Models for Single Cell Assays}
\description{
  MIMOSA implements mxitures of Dirichlet-multinomial or
  Beta-binomial models for paired count data from
  single--cell assays that typically arise in immunological
  studies. It can be used for ICS (Intracellular Cytokine
  Staining) assays to detect vaccine responders, for
  example, or to detect changes in proportions of cells
  expressing a gene, such as in Fluidigm Biomark
  Single--cell gene expression.
}
\references{
  Greg Finak, Andrew McDavid, Pratip Chattopadhyay, Maria
  Dominguez, Stephen C De Rosa, Mario Roederer, Raphael
  Gottardo Mixture Models for Single Cell Assays with
  Applications to Vaccine Studies Biostatistics, 2013,
  \url{http://biostatistics.oxfordjournals.org/content/early/2013/07/24/biostatistics.kxt024.abstract}
}

