\docType{package}
\name{MIMOSA-package}
\alias{MIMOSA-package}
\title{MIMOSA: Mixture Models for Single Cell Assays}
\description{
  MIMOSA implements mxitures of Dirichlet-multinomial or
  Beta-binomial models for paired count data from
  single--cell assays that typically arise in immunological
  studies. It can be used for ICS (Intracellular Cytokine
  Staining) assays to detect vaccine responders, for
  example, or to detect changes in proportions of cells
  expressing a gene, such as in Fluidigm Biomark
  Single--cell gene expression.
}
\references{
  Finak, Greg; McDavid, Andrew; Chattopadhyay, Pratip;
  Dominguez, Maria; De Rosa, Steve; Roederer, Mario;
  Gottardo, Raphael Mixture Models for Single Cell Assays
  with Applications to Vaccine Studies eprint
  arXiv:1208.5809 \url{http://arxiv.org/abs/1208.5809}
}

