% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{MIMOSAResult-class}
\alias{MIMOSAResult-class}
\title{Stores the result of a MIMOSA fitted model}
\description{
\code{MIMOSA} returns an object of \code{MIMOSAResult} irrespective of which
method / implementation is used to fit the data.
}

