\name{pData}
\alias{pData}
\alias{pData,MCMCResult-methods}
\alias{pData,MDMixResult-methods}
\alias{pData,MIMOSAResult-methods}
\alias{pData,MIMOSAResultList-methods}
\title{pData extract the phenoData table from a MIMOSA result}
\arguments{
  \item{object}{is the MIMOSAResult returned from a call to
  MIMOSA}
}
\value{
  an object of type \code{data.frame}
}
\description{
  pData extract the phenoData table from a MIMOSA result
}
\details{
  Extracts the phenoData data.frame from a MIMOSAResult
  object
}

