\name{print.MIMOSAResultList}
\alias{print.MIMOSAResultList}
\alias{show}
\alias{show,MIMOSAResult-method}
\title{Print a MIMOSAResultList}
\usage{
  \method{print}{MIMOSAResultList} (x, ...)
}
\arguments{
  \item{x}{a \code{MIMOSAResultList}}

  \item{...}{additional arguments passed down}
}
\description{
  Print a summary of the list of results returned by a call
  to \code{MIMOSA}
}

