% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R, R/shuffle_network.R
\name{randomize_net}
\alias{randomize_net}
\title{Randomize a metabolic network edge list by randomly sampling 2 edges and if it works, switching products}
\usage{
randomize_net(netw, n_reps)

randomize_net(netw, n_reps)
}
\arguments{
\item{netw}{Network edge list, format of the 3rd output item of generate_genomic_network}

\item{n_reps}{Number of successful edge switches to perform (default = 5000)}

\item{netw}{edge list network}

\item{n_reps}{number of times to successfully sample and switch}
}
\value{
Network edge list in the same format with the same compounds and degree distribution, but with randomized edges

randomized edge list
}
\description{
Randomly sample 2 edges and if it works, switch products
}
\examples{
randomize_net(edge_list, 3000)
randomize_net(allnet1, 1000)
}
