% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{run_all_metabolites}
\alias{run_all_metabolites}
\title{Runall function to do complete predictions and comparison for all shared metabolites}
\usage{
run_all_metabolites(genes, mets, file_prefix = "net1", correction = "fdr",
  cutoff = 0.1, net_method = "load", rxn_table_source = "", id_met = F,
  met_id_file = "", degree_filter = 0, minpath_file = "",
  cor_method = "spearman", net_file = "", nperm = 20000,
  nonzero_filter = 3, norm = T)
}
\arguments{
\item{genes}{Gene abundances}

\item{mets}{Metabolite abundances}

\item{file_prefix}{Prefix for output files}

\item{correction}{Type of multiple hypothesis correction to perform (bonferroni or fdr)}

\item{cutoff}{Q/P-value cutoff for significance}

\item{net_method}{Network generation method (see generate_genomic_network)}

\item{id_met}{Whether metabolites have putative identifications that need to be processed}

\item{met_id_file}{If id_met, file of metabolite identifications}

\item{degree_filter}{Threshold for filtering currency metabolites}

\item{minpath_file}{Optional file of pathways to filter network to}

\item{cor_method}{Either "spearman" or "pearson", default is Spearman}

\item{net_file}{file containing network template to use}

\item{nperm}{Number of permutations for Mantel test, default is 20000}

\item{nonzero_filter}{Minimum number of samples required to have nonzero concentrations and nonzero metabolic potential scores in order for metabolite to be evaluated, default is 3}

\item{norm}{Whether to normalize the network model coefficients by the total number of synthesis and degradation reactions for each metabolite}
}
\value{
No return, writes output to files.
}
\description{
Runall function to do complete predictions and comparison for all shared metabolites
}
\examples{

}
