% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{run_all_metabolites2}
\alias{run_all_metabolites2}
\title{Runall function to do complete predictions and evaluate classification of metabolites as high or low abundance}
\usage{
run_all_metabolites2(genes, mets, file_prefix = "net1", correction = "fdr",
  cutoff = 0.1, net_method = "load", rxn_table_source = "", id_met = F,
  met_id_file = "", degree_filter = 0, minpath_file = "", net_file = "",
  quant = 0.5, plot_rank = F, plot_continuous = F, nonzero_filter = 3,
  rel_abund_mets = F)
}
\arguments{
\item{genes}{Gene abundances}

\item{mets}{Metabolite abundances}

\item{file_prefix}{Prefix for output files}

\item{correction}{Type of multiple hypothesis correction to perform (bonferroni or fdr)}

\item{cutoff}{Q/P-value cutoff for significance}

\item{net_method}{Network generation method (see generate_genomic_network)}

\item{id_met}{Whether metabolites have putative identifications that need to be processed}

\item{met_id_file}{If id_met, file of metabolite identifications}

\item{degree_filter}{Threshold for filtering currency metabolites}

\item{minpath_file}{Optional file of pathways to filter network to}

\item{net_file}{file containing network template to use}

\item{quant}{Quantile above which a metabolite is "elevated", default is 0.5}

\item{plot_rank}{Whether to generate plots of ranks of metabolite concentrations and scores, default is F}

\item{plot_continuous}{Whether to generate plots of metabolite concentrations and scores, default is F}

\item{nonzero_filter}{Minimum number of samples required to have nonzero concentrations and nonzero metabolic potential scores in order for metabolite to be evaluated, default is 3}
}
\value{
No return, writes output to file
}
\description{
Runall function to do complete predictions and evaluate classification of metabolites as high or low abundance
}
\examples{

}

