% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleSpec_contributions.R
\name{single_spec_musicc}
\alias{single_spec_musicc}
\title{Get contributors for every metabolite and save}
\usage{
single_spec_musicc(contribs, otu_file)
}
\arguments{
\item{contribs}{data.table of OTU, genes, samples and abundances}

\item{otu_file}{File path to OTU abundance matrix}
}
\value{
contribs with abundances normalized to copy number
}
\description{
Get contributors for every metabolite and save
}

